+++ Runtime/Scripts/Interfaces/IMonitorHandle.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 868e0af3207148cf9358767acf579f60
-timeCreated: 1666470810
+fileFormatVersion: 2
+guid: 0f1797412cefe9242bd46631152c2cd3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
