--- Runtime/Scripts/Interfaces/IMonitorUnit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f1797412cefe9242bd46631152c2cd3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
